+++ iidm/iidm-impl/src/main/java/com/powsybl/iidm/network/impl/AbstractIdentifiable.java
@@ -33,7 +33,7 @@ abstract class AbstractIdentifiable<I extends Identifiable<I>> extends AbstractE
 
     private final Set<String> aliasesWithoutType = new HashSet<>();
     private final Map<String, String> aliasesByType = new HashMap<>();
-    private int sortIndex;
+    private long sortIndex;
 
     AbstractIdentifiable(String id, String name) {
         this.id = id;
@@ -221,12 +221,12 @@ abstract class AbstractIdentifiable<I extends Identifiable<I>> extends AbstractE
     }
 
     @Override
-    public int getSortIndex() {
+    public long getSortIndex() {
         return sortIndex;
     }
 
     @Override
-    public void setSortIndex(int sortIndex) {
+    public void setSortIndex(long sortIndex) {
         this.sortIndex = sortIndex;
     }
