+++ iidm/iidm-impl/src/main/java/com/powsybl/iidm/network/impl/NetworkIndex.java
@@ -23,7 +23,7 @@ class NetworkIndex {
 
     private final Map<String, Identifiable<?>> objectsById = new HashMap<>();
     private final Map<String, String> idByAlias = new HashMap<>();
-    private int sortIndex = 0;
+    private long sortIndex = 0;
 
     private final Map<Class<? extends Identifiable>, Set<Identifiable<?>>> objectsByClass = new HashMap<>();
